# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_message",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/groups": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Crée un nouveau groupe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "pattern": "^[^<>]*$",
                    "description": "Nom du groupe à créer",
                    "example": "Marketing team"
                  },
                  "description": {
                    "type": "string",
                    "pattern": "^[^<>]*$",
                    "description": "Description du groupe à créer",
                    "example": "Group of the marketing team"
                  },
                  "memberIds": {
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        },
                        {
                          "type": "string",
                          "pattern": "^(?:user-)?\\d+$"
                        }
                      ],
                      "description": "Identifiant (user-<id> ou entier)",
                      "example": 3
                    },
                    "description": "Liste des identifiants des membres à ajouter au groupe"
                  }
                },
                "required": [
                  "name",
                  "memberIds"
                ],
                "description": "Corps de la requête pour créer un nouveau groupe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Groupe créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conversation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            }
                          ],
                          "description": "Identifiant unique, peut être une chaîne ou un nombre",
                          "example": "12345"
                        },
                        "name": {
                          "type": "string",
                          "description": "Nom de la conversation",
                          "example": "Discussion de groupe"
                        },
                        "department": {
                          "type": "string",
                          "description": "Département associé à la conversation",
                          "example": "Marketing"
                        },
                        "kind": {
                          "type": "string",
                          "enum": [
                            "direct",
                            "group"
                          ],
                          "description": "Conversation type: `direct` for a conversation created by `POST /direct-messages` between the caller and one contact, `group` otherwise",
                          "example": "direct"
                        },
                        "contactId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            }
                          ],
                          "description": "Direct conversations only: id of the other participant (the contact), to post a new message to that contact in this conversation instead of creating another one",
                          "example": "user-8"
                        },
                        "avatarUrl": {
                          "type": "string",
                          "format": "uri",
                          "description": "URL de l’avatar de la conversation",
                          "example": "https://example.com/avatar.png"
                        },
                        "initials": {
                          "type": "string",
                          "description": "Initiales de la conversation",
                          "example": "DG"
                        },
                        "presence": {
                          "type": "string",
                          "enum": [
                            "online",
                            "offline",
                            "away"
                          ],
                          "description": "Présence de l’utilisateur, en ligne, hors ligne ou absent",
                          "example": "online"
                        },
                        "lastMessage": {
                          "type": "string",
                          "description": "Dernier message de la conversation",
                          "example": "Bonjour à tous !"
                        },
                        "lastMessageAt": {
                          "type": "string",
                          "description": "Date et heure du dernier message (format ISO)",
                          "example": "2026-06-23T12:32:00Z"
                        },
                        "unreadCount": {
                          "type": "integer",
                          "description": "Nombre de messages non lus dans la conversation",
                          "example": 5
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "description": "Détails du groupe créé"
                    }
                  },
                  "required": [
                    "conversation"
                  ],
                  "description": "Réponse contenant les détails du groupe créé"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body (details lists the invalid fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Message API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}